Abstract Facepile component
