+++ mobile/components/ThreadItem/style.js
@@ -1,7 +1,6 @@
 // @flow
-import styled, { css } from 'styled-components/native';
-import { TouchableOpacity } from 'react-native';
-import Avatar from '../Avatar';
+import styled from 'styled-components/native';
+import TouchableOpacity from '../TouchableOpacity';
 
 export const InboxThreadContent = styled.View``;
 
@@ -41,24 +40,6 @@ export const MetaTextPill = styled(MetaText)`
   overflow: hidden;
 `;
 
-export const FacepileContainer = styled.View`
-  display: flex;
-  flex-direction: row;
-`;
-
-export const EmptyParticipantHead = styled.Text`
-  color: ${props => props.theme.text.alt};
-  background: ${props => props.theme.bg.wash};
-  border-radius: ${props => (props.radius ? `${props.radius}px` : '15px')};
-  text-align: center;
-  text-align-vertical: center;
-  font-size: 12px;
-  font-weight: 600;
-  height: ${props => (props.size ? `${props.size}px` : '30px')};
-  width: ${props => (props.size ? `${props.size}px` : '30px')};
-  overflow: hidden;
-`;
-
 export const ThreadCommunityInfoWrapper = styled.View`
   display: flex;
   flex-direction: row;
@@ -86,18 +67,4 @@ export const ThreadChannelPill = styled(TouchableOpacity)`
 export const ThreadChannelName = styled.Text`
   color: ${props => props.theme.text.alt};
   font-size: 12px;
-`;
-
-const stackingStyles = css`
-  margin-right: -10px;
-  border-width: 2px;
-  border-color: #ffffff;
-`;
-
-export const StackedEmptyParticipantHead = styled(EmptyParticipantHead)`
-  ${stackingStyles};
-`;
-
-export const StackedAvatar = styled(Avatar)`
-  ${stackingStyles};
 `;
